refactor: relocate VersionedEnvelope in prep for isolation
