+++ sawtooth-daml-common/src/main/java/com/blockchaintp/utils/VersionedEnvelopeUtils.java
@@ -1,185 +1,33 @@
-/*
- * Copyright 2019 Blockchain Technology Partners Licensed under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance with the License. You may obtain a
- * copy of the License at http://www.apache.org/licenses/LICENSE-2.0 Unless required by applicable
- * law or agreed to in writing, software distributed under the License is distributed on an "AS IS"
- * BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License
- * for the specific language governing permissions and limitations under the License.
- * ------------------------------------------------------------------------------
- */
+
 package com.blockchaintp.utils;
 
-import static org.bitcoinj.core.Utils.HEX;
 import static sawtooth.sdk.processor.Utils.hash512;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-import java.security.SecureRandom;
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.List;
 import java.util.zip.DataFormatException;
 import java.util.zip.Deflater;
 import java.util.zip.Inflater;
 
-import com.blockchaintp.utils.protobuf.VersionedEnvelope;
-import com.blockchaintp.keymanager.KeyManager;
 import com.google.protobuf.ByteString;
 import com.google.protobuf.InvalidProtocolBufferException;
+import com.blockchaintp.utils.protobuf.VersionedEnvelope;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.spongycastle.util.Arrays;
 
-import sawtooth.sdk.messaging.Future;
-import sawtooth.sdk.messaging.Stream;
-import sawtooth.sdk.protobuf.Batch;
-import sawtooth.sdk.protobuf.BatchHeader;
-import sawtooth.sdk.protobuf.ClientBatchSubmitRequest;
-import sawtooth.sdk.protobuf.Message;
-import sawtooth.sdk.protobuf.Transaction;
-import sawtooth.sdk.protobuf.TransactionHeader;
-
-/**
- * Utility methods and constants useful for interacting with Sawtooth as a
- * client.
- */
-public final class SawtoothClientUtils {
+public class VersionedEnvelopeUtils {
 
   private static final int COMPRESS_BUFFER_SIZE = 1024;
 
-  private SawtoothClientUtils() {
-  }
+  private static final Logger LOGGER = LoggerFactory.getLogger(VersionedEnvelopeUtils.class);
 
-  /**
-   * A SecureRandom for use in creating batches.
-   */
-  private static SecureRandom secureRandom;
-  private static int randomBytesGenerated;
+  private VersionedEnvelopeUtils() {}
 
-  private static final Logger LOGGER = LoggerFactory.getLogger(SawtoothClientUtils.class);
-
-  private static final int MAX_BYTES_PER_SEED = 10 * 1024 * 1024;
-
-  /**
-   * Create a batch. Transaction passed are specified to be processed in order.
-   *
-   * @param keyManager A key manager handling the identity to be used for signing
-   *                   etc.
-   * @param txns       a collection of transactions
-   * @return the assembled batch
-   */
-  public static Batch makeSawtoothBatch(final KeyManager keyManager, final Collection<Transaction> txns) {
-    BatchHeader.Builder batchHeaderBldr = BatchHeader.newBuilder().clearSignerPublicKey()
-        .setSignerPublicKey(keyManager.getPublicKeyInHex());
-    for (Transaction tx : txns) {
-      batchHeaderBldr.addTransactionIds(tx.getHeaderSignature());
-    }
-    var batchHeader = batchHeaderBldr.build();
-    String signedHeader = keyManager.sign(batchHeader.toByteArray());
-    return Batch.newBuilder().setHeader(batchHeader.toByteString()).setHeaderSignature(signedHeader)
-        .addAllTransactions(txns).build();
-  }
-
-  /**
-   * Make a sawtooth transaction based on the provided parameters.
-   *
-   * @param keyManager              a keymanager to provide the necessary identity
-   *                                info
-   * @param familyName              the family name
-   * @param familyVersion           the family version
-   * @param inputAddresses          state addresses required for input to this
-   *                                transaction
-   * @param outputAddresses         state addresses this transaction will output
-   *                                to
-   * @param dependentTransactionIds transaction ids this transaction is dependent
-   *                                upon
-   * @param payload                 the ByteString payload of this transaction
-   * @return the transaction
-   */
-  public static Transaction makeSawtoothTransaction(final KeyManager keyManager, final String familyName,
-      final String familyVersion, final Collection<String> inputAddresses, final Collection<String> outputAddresses,
-      final Collection<String> dependentTransactionIds, final ByteString payload) {
-    ByteString wrappedPayload = SawtoothClientUtils.wrap(payload);
-    String payloadHash = getHash(wrappedPayload);
-    TransactionHeader.Builder txnHeaderBldr = TransactionHeader.newBuilder().setFamilyName(familyName)
-        .setFamilyVersion(familyVersion).clearBatcherPublicKey().setBatcherPublicKey(keyManager.getPublicKeyInHex())
-        .setNonce(SawtoothClientUtils.generateNonce()).setPayloadSha512(payloadHash).addAllInputs(inputAddresses)
-        .addAllOutputs(outputAddresses).addAllDependencies(dependentTransactionIds)
-        .setSignerPublicKey(keyManager.getPublicKeyInHex());
-    TransactionHeader txnHeader = txnHeaderBldr.build();
-
-    String signedHeader = keyManager.sign(txnHeader.toByteArray());
-    return Transaction.newBuilder().setHeader(txnHeader.toByteString()).setHeaderSignature(signedHeader)
-        .setPayload(wrappedPayload).build();
-  }
-
-  /**
-   * Generate a random nonce.
-   *
-   * @return the nonce
-   */
-  public static String generateNonce() {
-    final var seedByteCount = 20;
-    synchronized (SawtoothClientUtils.class) {
-      if (null == secureRandom) {
-        LOGGER.debug("Generating nonce - acquiring SecureRandom");
-        secureRandom = new SecureRandom();
-        randomBytesGenerated = -1;
-      }
-      if (randomBytesGenerated == -1) {
-        LOGGER.debug("Generating nonce - regenerating seed");
-        secureRandom.setSeed(secureRandom.generateSeed(seedByteCount));
-        randomBytesGenerated = seedByteCount;
-      } else {
-        randomBytesGenerated += seedByteCount;
-      }
-      if (randomBytesGenerated > MAX_BYTES_PER_SEED) {
-        randomBytesGenerated = -1;
-      }
-    }
-
-    var nonceBytes = new byte[seedByteCount];
-    LOGGER.debug("Generating nonce - generating nonce");
-    secureRandom.nextBytes(nonceBytes);
-    LOGGER.debug("Generating nonce - nonce generated");
-    return HEX.encode(nonceBytes);
-  }
-
-  /**
-   * For a given string return its sha512, transform the encoding problems into
-   * RuntimeErrors.
-   *
-   * @param arg a string
-   * @return the hash512 of the string
-   */
-  public static String getHash(final String arg) {
-    byte[] data = arg.getBytes(StandardCharsets.UTF_8);
-    return getHash(data);
-  }
-
-  /**
-   * For a given ByteString return its sha512.
-   *
-   * @param arg a ByteString
-   * @return the hash512 of the ByteString
-   */
-  public static String getHash(final ByteString arg) {
-    return getHash(arg.toByteArray());
-  }
-
-  /**
-   * For a given array of bytes return its sha512.
-   *
-   * @param arg the bytes
-   * @return the hash512 of the byte array
-   */
-  public static String getHash(final byte[] arg) {
-    return hash512(arg);
-  }
-
-  /**
+    /**
    * Wrap the given ByteString with a VersionEnvelope and return.
    *
    * @param value The value to be wrapped
@@ -188,7 +36,7 @@ public final class SawtoothClientUtils {
    */
   public static ByteString wrap(final ByteString value) throws InternalError {
     var data = compressByteString(value);
-    String contentHash = getHash(data.toByteArray());
+    String contentHash = hash512(data.toByteArray());
     return VersionedEnvelope.newBuilder().setData(data).setParts(1).setPartNumber(0).setContentHash(contentHash).build()
         .toByteString();
   }
@@ -233,7 +81,7 @@ public final class SawtoothClientUtils {
    */
   public static List<ByteString> wrapMultipart(final ByteString value, final int maxPartSize) throws InternalError {
     var compressedData = compressByteString(value);
-    String contentHash = getHash(compressedData.toByteArray());
+    String contentHash = hash512(compressedData.toByteArray());
     List<ByteString> retList = new ArrayList<>();
     byte[] compressedBytes = compressedData.toByteArray();
     List<byte[]> splitBytes = splitBytes(compressedBytes, maxPartSize);
@@ -261,7 +109,7 @@ public final class SawtoothClientUtils {
       accumulatedBytes = Arrays.concatenate(accumulatedBytes, e.getData().toByteArray());
     }
     var data = ByteString.copyFrom(accumulatedBytes);
-    String assembledHash = getHash(accumulatedBytes);
+    String assembledHash = hash512(accumulatedBytes);
     if (assembledHash.equals(contentHash)) {
       LOGGER.trace("Assembled hash looks good {} = {}", contentHash, assembledHash);
     } else {
@@ -286,7 +134,7 @@ public final class SawtoothClientUtils {
       switch (envelope.getVersion()) {
         case "":
         case "1":
-          String assembledHash = getHash(envelope.getData().toByteArray());
+          String assembledHash = hash512(envelope.getData().toByteArray());
           String contentHash = envelope.getContentHash();
           if (contentHash.equals(assembledHash)) {
             LOGGER.trace("Assembled hash looks good {} = {}", contentHash, assembledHash);
@@ -366,15 +214,4 @@ public final class SawtoothClientUtils {
     }
   }
 
-  /**
-   * Subit the Batch to the sawtooth Stream.
-   *
-   * @param batch  the batch to submit
-   * @param stream the stream to submit to
-   * @return a Future upon which to wait
-   */
-  public static Future submitBatch(final Batch batch, final Stream stream) {
-    final ClientBatchSubmitRequest cbsReq = ClientBatchSubmitRequest.newBuilder().addBatches(batch).build();
-    return stream.send(Message.MessageType.CLIENT_BATCH_SUBMIT_REQUEST, cbsReq.toByteString());
-  }
 }
