+++ sawtooth-daml-rpc/pom.xml
@@ -49,19 +49,6 @@
             <manifest>
               <addClasspath>true</addClasspath>
               <classpathPrefix>lib/</classpathPrefix>
-              <mainClass>com.blockchaintp.sawtooth.daml.rpc.JwtGenerator</mainClass>
-            </manifest>
-          </archive>
-        </configuration>
-      </plugin>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-jar-plugin</artifactId>
-        <configuration>
-          <archive>
-            <manifest>
-              <addClasspath>true</addClasspath>
-              <classpathPrefix>lib/</classpathPrefix>
               <mainClass>com.blockchaintp.sawtooth.daml.rpc.Main</mainClass>
             </manifest>
           </archive>
